Extract Payments classes out of the api library
