+++ libs/api/src/lib/graphql/paymentMethod.ts
@@ -8,7 +8,7 @@ import {
 } from 'graphql'
 import {Context} from '../context'
 import {GraphQLDateTime} from 'graphql-scalars'
-import {PaymentProvider} from '../payments/paymentProvider'
+import {PaymentProvider} from '@wepublish/payments'
 import {createProxyingResolver} from '../utility'
 import {GraphQLSlug} from './slug'
 import {PaymentMethod} from '@prisma/client'
