+++ libs/api/src/lib/memberContext.ts
@@ -21,7 +21,7 @@ import {MemberPlanWithPaymentMethods} from './db/memberPlan'
 import {SubscriptionWithRelations} from './db/subscription'
 import {InternalError, NotFound, PaymentConfigurationNotAllowed, UserInputError} from './error'
 import {MailContext} from '@wepublish/mails'
-import {PaymentProvider} from './payments/paymentProvider'
+import {PaymentProvider} from '@wepublish/payments'
 import {logger} from '@wepublish/utils'
 import {ONE_DAY_IN_MILLISECONDS, ONE_MONTH_IN_MILLISECONDS} from './utility'
 import {SubscriptionEventDictionary} from '@wepublish/membership/api'
