+++ libs/payments/README.md
@@ -1,0 +1,7 @@
+# payments
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test payments` to execute the unit tests via [Jest](https://jestjs.io).
