+++ libs/payments/src/lib/payment-provider/index.ts
@@ -1,0 +1,5 @@
+export * from './paymentProvider'
+export * from './payrexxPaymentProvider'
+export * from './payrexxSubscriptionPaymentProvider'
+export * from './stripePaymentProvider'
+export * from './stripeCheckoutPaymentProvider'
